Add git_is_tag function and its appropriate tests
The git_is_tag functions checks if the HEAD ref points to a tag; it can
be a simple, annotated or signed tag.

When checking out a tag, you enter in "detached head" mode, for that
reason I'm using the git_is_detached_head function in here.

This is useful in use with the git_branch_name function, which returns
the tag name if HEAD points to one. I'm using it on my prompt to
display a diff icon when on top of a tag: https://git.io/vX4Ku