Make pattern matching of git_ahead stricter
This to prevent earlier return and false results.

For example, I had the following state on a repo:

```shell
$ git rev-list --count --left-right "@{upstream}...HEAD"
10  36
```

And git_ahead didn't report "diverged" correctly:

```shell
$ git_ahead
+
```

Reporting "ahead" instead because the of trailing zero of 10 on the
left column matched the "/[0-9]+\t0/" pattern.

Making the pattern stricter using the caret to match the start of the
line fixes such issue.